Create README.md
This is the backend of an ecommerce website. I'm slowly but surely building out everything needed for an ecommerce website to run using the Django framework.
This repo is a new repo since beginning to test out Workers and Message Queues using Redis and Celery. I was forced to move into a Linux environment use WSL 2 as Celery was no longer compatible with Windows.
Caused me a hell of a lot of problems trying to get it up and running with version control issues across various tools in the stack. Nightmare actually for a first timer solving these issues. Got there in the end though and happy about the learning experience as a result.
Onwards!